// institutions/NeispuoSOFront/src/app/enums/fieldType.enum.ts
export enum FieldType {
  Input = "input",
  Button = "button",
  Select = "select",
  Date = "date",
  Radiobutton = "radiobutton",
  Checkbox = "checkbox",
  Searchselect = "searchselect",
  Multiselect = "multiselect",
  Label = "label",
  Info = "info",
  Breakpoint = "breakpoint",
  DateTime = "datetime"
}